+++ src/IconFactory/IconFactory.tsx
@@ -5,6 +5,7 @@ import AddProductIcon from '../../assets/icons/mono-color/add-product.svg'
 import BulkProductIcon from '../../assets/icons/mono-color/bulk-product.svg'
 import CompositeProductIcon from '../../assets/icons/mono-color/composite-product.svg'
 import SingleProductIcon from '../../assets/icons/mono-color/single-product.svg'
+import TerminalIcon from '../../assets/icons/mono-color/terminal.svg'
 // System Icons
 import ArrowRight from '../../assets/icons/system/arrow-right.svg'
 import Card from '../../assets/icons/system/card.svg'
@@ -12,6 +13,7 @@ import ChevronDown from '../../assets/icons/system/chevron-down.svg'
 import ChevronLeft from '../../assets/icons/system/chevron-left.svg'
 import ChevronRight from '../../assets/icons/system/chevron-right.svg'
 import ChevronUp from '../../assets/icons/system/chevron-up.svg'
+import Clock from '../../assets/icons/system/clock.svg'
 import Close from '../../assets/icons/system/close.svg'
 import CloudUpload from '../../assets/icons/system/cloud-upload.svg'
 import Data from '../../assets/icons/system/data.svg'
@@ -24,6 +26,7 @@ import Eye from '../../assets/icons/system/eye.svg'
 import File from '../../assets/icons/system/file.svg'
 import Filter from '../../assets/icons/system/filter.svg'
 import Home from '../../assets/icons/system/home.svg'
+import Indicator from '../../assets/icons/system/indicator.svg'
 import InfoCircle from '../../assets/icons/system/info-circle.svg'
 import LongArrowRight from '../../assets/icons/system/long-arrow.svg'
 import Medal from '../../assets/icons/system/medal.svg'
@@ -32,6 +35,7 @@ import Notification from '../../assets/icons/system/notification.svg'
 import People from '../../assets/icons/system/people.svg'
 import Plus from '../../assets/icons/system/plus.svg'
 import Receipt from '../../assets/icons/system/receipt.svg'
+import Scan from '../../assets/icons/system/scan.svg'
 import Search from '../../assets/icons/system/search.svg'
 import Setting from '../../assets/icons/system/setting.svg'
 import ShopCart from '../../assets/icons/system/shop-cart.svg'
@@ -47,6 +51,7 @@ export type IconName =
   | 'chevron-right'
   | 'chevron-left'
   | 'close'
+  | 'clock'
   | 'data'
   | 'edit'
   | 'eye-closed'
@@ -54,6 +59,7 @@ export type IconName =
   | 'file'
   | 'filter'
   | 'home'
+  | 'indicator'
   | 'info-circle'
   | 'medal'
   | 'note'
@@ -64,6 +70,7 @@ export type IconName =
   | 'search'
   | 'setting'
   | 'shop-cart'
+  | 'scan'
   | 'trash'
   | 'user-circle'
   | 'ellipsis'
@@ -72,6 +79,7 @@ export type IconName =
   | 'bulk-product'
   | 'composite-product'
   | 'add-product'
+  | 'terminal'
   | 'long-arrow-right'
   | 'download'
   | 'cloud-upload'
@@ -85,6 +93,7 @@ const iconMap: Record<IconName, FC<SVGProps<SVGSVGElement>>> = {
   'chevron-right': ChevronRight,
   'chevron-left': ChevronLeft,
   close: Close,
+  clock: Clock,
   data: Data,
   edit: Edit,
   'eye-closed': EyeClosed,
@@ -92,6 +101,7 @@ const iconMap: Record<IconName, FC<SVGProps<SVGSVGElement>>> = {
   file: File,
   filter: Filter,
   home: Home,
+  indicator: Indicator,
   'info-circle': InfoCircle,
   medal: Medal,
   note: Note,
@@ -102,6 +112,7 @@ const iconMap: Record<IconName, FC<SVGProps<SVGSVGElement>>> = {
   search: Search,
   setting: Setting,
   'shop-cart': ShopCart,
+  scan: Scan,
   trash: Trash,
   'user-circle': UserCircle,
   ellipsis: Ellipsis,
@@ -110,6 +121,7 @@ const iconMap: Record<IconName, FC<SVGProps<SVGSVGElement>>> = {
   'composite-product': CompositeProductIcon,
   'bulk-product': BulkProductIcon,
   'add-product': AddProductIcon,
+  terminal: TerminalIcon,
   'long-arrow-right': LongArrowRight,
   download: Download,
   'cloud-upload': CloudUpload,